pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "image-preprocess-api"
version = "0.1.0"
description = "API service for image and dataset preprocessing"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "self_api team" }
]
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
  "pillow>=10.4.0,<12.0.0",
  "pydantic-settings>=2.4.0,<3.0.0",
  "python-multipart>=0.0.9,<1.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0,<9.0.0",
  "httpx>=0.27.0,<1.0.0"
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
